single float literals seem to work

+++ example/12_business_card_ray/source/nvcut_00_cuda_main.cu
@@ -37,13 +37,13 @@ struct v {
     z = c;
   };
   __device__ v operator!() {
-    return ((*this) * ((((1.0)) / (sqrt(*this % *this)))));
+    return ((*this) * ((((1.0f)) / (sqrt(*this % *this)))));
   };
 };
 __device__ int g_seed = 1;
 __device__ float R() {
   g_seed = ((((214013) * (g_seed))) + (2531011));
-  return (((((g_seed) >> (16)) & (0x7fff))) / ((66635.)));
+  return (((((g_seed) >> (16)) & (0x7fff))) / ((66635.f)));
 };
 __device__ int G[9] = {247570, 280596, 280600, 249748, 18578,
                        18577,  231184, 16,     16};
@@ -61,7 +61,7 @@ __device__ int TraceRay(v origin, v destination, float &tau, v &normal) {
       if ((((G[j]) & (1))) << (k)) {
         auto p = ((origin) + (v(-k, 0, ((-j) - (4)))));
         auto b = p % destination;
-        auto c = ((p % p) - ((1.0)));
+        auto c = ((p % p) - ((1.0f)));
         auto q = ((((b) * (b))) - (c));
         if (0 < q) {
           auto s = ((-b) - (sqrt(q)));
@@ -77,42 +77,42 @@ __device__ int TraceRay(v origin, v destination, float &tau, v &normal) {
   return m;
 };
 __device__ v Sample(v origin, v destination, int r) {
-  auto tau = (0.);
+  auto tau = (0.f);
   auto normal = v();
   if (4 < r) {
     return v();
   };
   auto match = TraceRay(origin, destination, tau, normal);
   if (!(match)) {
-    return ((v((0.70), (0.60), 1)) * (pow(((1) - (destination.z)), 4)));
+    return ((v((0.70f), (0.60f), 1)) * (powf(((1) - (destination.z)), 4)));
   };
   auto intersection = ((origin) + (((destination) * (tau))));
   auto light_dir =
       ((!(v(((9) + (R())), ((9) + (R())), 16))) + (((intersection) * (-1))));
   auto half_vec =
       ((destination) + (((normal) * (((normal % destination) * (2))))));
-  auto color = (1.0);
+  auto color = (1.0f);
   return v(((10) * (origin.x)), color, color);
 };
 __global__ void GetColor(unsigned char *img) {
   auto x = blockIdx.x;
   auto y = threadIdx.x;
-  auto cam_dir = !(v((-6.0), (-16.), (0.)));
+  auto cam_dir = !(v((-6.0f), (-16.f), (0.f)));
   auto s = (2.00e-3);
-  auto cam_up = ((!(((v((0.), (0.), (1.0))) ^ (cam_dir)))) * (s));
+  auto cam_up = ((!(((v((0.f), (0.f), (1.0f))) ^ (cam_dir)))) * (s));
   auto cam_right = ((!(((cam_dir) ^ (cam_up)))) * (s));
   auto eye_offset = ((((((cam_up) + (cam_right))) * (-256))) + (cam_dir));
-  auto color = v((13.), (13.), (13.));
+  auto color = v((13.f), (13.f), (13.f));
   for (int r = 0; r < 64; (r) += (1)) {
-    auto delta = ((((cam_up) * (((R()) - ((0.50)))) * (99))) +
-                  (((cam_right) * (((R()) - ((0.50)))) * (99))));
+    auto delta = ((((cam_up) * (((R()) - ((0.50f)))) * (99))) +
+                  (((cam_right) * (((R()) - ((0.50f)))) * (99))));
     color =
         ((((Sample(((v(17, 16, 8)) + (delta)),
                    !(((((((delta) * (-1))) + (((cam_up) * (((R()) + (x))))) +
                         (((cam_right) * (((y) + (R()))))) + (eye_offset))) *
                       (16))),
                    0)) *
-           ((3.50)))) +
+           ((3.50f)))) +
          (color));
   }
   img[((((DIM) * (y) * (BPP))) + (((BPP) * (x))) + (0))] = color.x;
